Working out some issues with code timer
intellisense just breaks because of the typedef struct name and I dont know why. Basic forking should be working but I wont know until he grades it because he gave no feedback on the last round. A few updates to Process Affinity to start to applying the new code timer functionality to it.