Basic functionality working
Able to pass a string to the thread that can be interacted with and returns an int that can be interacted with. Next is making the input value a struct to contain the char and the string. Timers will be applied to the switch case for the full time and internally for the individual thread time.