struct passing
struct can now be passed into thread, just needs to be edited and have the value counted. Bug with the input makes the string for the search char also attach the string input for the pre parsed thread number. It might be because I have the char array that its looking at set to 1 when it should be zero.